Move Store into the data module
